+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [0.1.7] - 2018-06-13
 ### Added
 - Displays property type when inherited from referenced schema
+- Creates "Overview" and "Remarks" headers for human-written sections
 
 ## [0.1.7] - 2018-06-13
 ### Fixed